IPYNB File
This notebook consists of the general procedure to perform analysis and modelling or forecasting on any time series data. The data used is Air passengers and the aim of the notebook is to build a model to forecast how many numbers of passengers are going to abort these airlines at the month level in the future( next 10 years). I had build a ARIMA model to forecast future values.